+++ ui/src/app/core/manage-plugins/plugin-config/plugin-config.component.ts
@@ -24,8 +24,8 @@ import { SchemaFormComponent } from '@/app/core/components/schema-form/schema-fo
 import { PluginsMarkdownDirective } from '@/app/core/directives/plugins.markdown.directive'
 import { HomebridgeDeconzComponent } from '@/app/core/manage-plugins/custom-plugins/homebridge-deconz/homebridge-deconz.component'
 import { HomebridgeHueComponent } from '@/app/core/manage-plugins/custom-plugins/homebridge-hue/homebridge-hue.component'
-import { InterpolateMdPipe } from '@/app/core/manage-plugins/interpolate-md.pipe'
 import { ManagePluginsService } from '@/app/core/manage-plugins/manage-plugins.service'
+import { InterpolateMdPipe } from '@/app/core/pipes/interpolate-md.pipe'
 import { SettingsService } from '@/app/core/settings.service'
 
 export interface PluginConfigBlock {