+++ ui/src/app/modules/platform-tools/platform-tools.module.ts
@@ -1,4 +1,3 @@
-import { CommonModule } from '@angular/common'
 import { NgModule } from '@angular/core'
 
 import { PlatformToolsRoutingModule } from '@/app/modules/platform-tools/platform-tools-routing.module'
@@ -6,7 +5,6 @@ import { PlatformToolsRoutingModule } from '@/app/modules/platform-tools/platfor
 @NgModule({
   declarations: [],
   imports: [
-    CommonModule,
     PlatformToolsRoutingModule,
   ],
 })