+++ ui/src/app/modules/power-options/power-options.module.ts
@@ -1,20 +1,10 @@
-import { CommonModule } from '@angular/common'
 import { NgModule } from '@angular/core'
-import { ReactiveFormsModule } from '@angular/forms'
-import { NgbModule } from '@ng-bootstrap/ng-bootstrap'
-import { TranslateModule } from '@ngx-translate/core'
 
 import { PowerOptionsRoutingModule } from '@/app/modules/power-options/power-options-routing.module'
-import { PowerOptionsComponent } from '@/app/modules/power-options/power-options.component'
 
 @NgModule({
   imports: [
-    CommonModule,
     PowerOptionsRoutingModule,
-    NgbModule,
-    ReactiveFormsModule,
-    TranslateModule,
-    PowerOptionsComponent,
   ],
 })
 export class PowerOptionsModule {}