+++ ui/src/app/modules/setup-wizard/setup-wizard.module.ts
@@ -1,22 +1,11 @@
-import { CommonModule } from '@angular/common'
 import { NgModule } from '@angular/core'
-import { FormsModule, ReactiveFormsModule } from '@angular/forms'
-import { NgbModule } from '@ng-bootstrap/ng-bootstrap'
-import { TranslateModule } from '@ngx-translate/core'
 
 import { SetupWizardRoutingModule } from '@/app/modules/setup-wizard/setup-wizard-routing.module'
-import { SetupWizardComponent } from '@/app/modules/setup-wizard/setup-wizard.component'
 import { SetupWizardGuard } from '@/app/modules/setup-wizard/setup-wizard.guard'
 
 @NgModule({
   imports: [
-    CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    TranslateModule,
-    NgbModule,
     SetupWizardRoutingModule,
-    SetupWizardComponent,
   ],
   providers: [
     SetupWizardGuard,